backend/requirements.txt
asgiref==3.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.6.15
chardet==3.0.4
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
ffmpeg==1.4
filelock==3.18.0
fsspec==2025.7.0
gunicorn==23.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
idna==2.10
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
packaging==25.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydub==0.25.1
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rfc3986==1.5.0
rpds-py==0.26.0
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0